backend/requirements.txt
fastapi
uvicorn
aiokafka
asyncpg
sqlalchemy
pydantic
python-dotenv
google-generativeai